Fix named clone hats not triggering
This also includes a major refactor to all named clone logic. In the end
it turned out to be better: Got rid of a dirty hack and it actually works.

To summarize:
* RenderedTargets have a cloneName property, which is null when they are
the original sprint or an anonymous (classical) clone.
* If the cloneName property is not null, startHats for named clones
blocks are called.
* RenderedTarget.makeClone() gained a parameter, namely the cloneName.
It is null if the clone is anonymous, and contains the clone name as a
string if the clone is named. The cloneName property of the
new clone is set according to this parameter.
* Got rid of the clone name placeholder hack.
* Got rid of the separate createNamedClone() opcode function. It
duplicated a lot of code with createClone(), the functionality is merged
in createClone() now.

WIP #34