Use '➡' as the separator in dict's array representation
This further extends the hack that I have made in order to support the
dict iterator block. Basically, a '➡' character is placed between the
key/value pair in a string, so one can split the string on that character
to separate the key and value. Who in their right mind will use the '➡'
character in a dict key or value, anyway?

Return an array of strings (of key/value pairs separated by '➡') to
the GUI, so that the dict monitor can split the strings on it and
separate the keys and values.